Feat/permissions

+++ app/Http/Controllers/ConcernController.php
@@ -38,6 +38,9 @@ class ConcernController extends Controller
      */
     public function index()
     {
+        if (auth()->user()->cannot('view-all', $this->concern)) {
+            return redirect('home')->with('alert.danger', 'You do not have access to this page.');
+        }
         $concerns = $this->concern->with([
             'user:id,name',
             'students:student_id,forename,surname,year_group',
@@ -53,6 +56,10 @@ class ConcernController extends Controller
      */
     public function create()
     {
+        if (auth()->user()->cannot('create', $this->concern)) {
+            return redirect('home')->with('alert.danger', 'You do not have access to this page.');
+        }
+
         return view('concerns.create', [
             'groups' => $this->group->all(),
             'students' => $this->student->all()
@@ -67,6 +74,10 @@ class ConcernController extends Controller
      */
     public function store(ConcernRequest $request)
     {
+        if (auth()->user()->cannot('create', $this->concern)) {
+            return redirect('home')->with('alert.danger', 'You do not have access to this page.');
+        }
+
         $concern = $this->concern->create([
             'user_id' => $request->user_id,
             'title' => $request->title,
@@ -100,6 +111,10 @@ class ConcernController extends Controller
      */
     public function show(Concern $concern)
     {
+        if (auth()->user()->cannot('view', $concern)) {
+            return back()->with('alert.danger', 'You do not have access to view this concern.');
+        }
+
         $concern = $this->concern->with([
             'user:id,name',
             'students:student_id,forename,surname,year_group',
@@ -112,24 +127,13 @@ class ConcernController extends Controller
         return view('concerns.show', ['concern' => $concern]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit($id)
     {
         // TODO: Create the form for editing concerns
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
         // TODO: Create the method to update the concern
@@ -143,6 +147,10 @@ class ConcernController extends Controller
      */
     public function destroy(Concern $id)
     {
+        if (auth()->user()->cannot('delete', $this->concern)) {
+            return redirect('home')->with('alert.danger', 'You do not have access to this page.');
+        }
+
         $this->concern->destroy($id);
     }
 }
